Hidden skills enhancement & bug fix.
Only display the hidden-skills options if there are hidden skills in our
filtered list (so, if there are hidden skills which we weren't going to
show anyway, we hide the button).

This also fixes a crash when you mark all filtered skills as hidden, and
are not displaying hidden skills.